git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@4674 3895a583-e809-0410-9bb6-aa9463026377

+++ MSH/msh_elem.h
@@ -97,12 +97,12 @@ public:
 	}
 
 	void GetNodes(vec<CNode*>& ele_nodes) {
-		for (size_t i=0; i<nodes.Size(); i++)
+		for (int i=0; i<nodes.Size(); i++)
 			ele_nodes[i] = nodes[i];
 	}
 
 	void GetNodes(std::vector<CNode*>& nodesVec) {
-		for (size_t i=0; i<nodes.Size(); i++)
+		for (int i=0; i<nodes.Size(); i++)
 			nodesVec.push_back(nodes[i]);
 	}
